refactor on RestAuthApiClient
