+++ src/lib/api/RestAuthApiClient/index.ts
@@ -1,0 +1,2 @@
+export * from "./RestAuthApiClient";
+export * from "./types";
